requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.3
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.0
backoff==2.2.1
bitsandbytes==0.42.0
Brotli==1.0.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
cmake==4.0.3
comm==0.2.1
contourpy==1.3.1
cryptography==45.0.4
cycler==0.11.0
Cython==3.1.2
debugpy==1.8.11
decorator==5.1.1
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
executing==0.8.3
fastmcp==2.10.0
filelock==3.18.0
fonttools==4.55.3
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.0
idna==3.7
inflect==7.5.0
ipykernel==6.29.5
ipython==8.37.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mcp==1.10.1
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.2.6
omegaconf==2.3.0
openai==1.86.0
openapi-pydantic==0.5.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.0
parso==0.8.4
pexpect==4.8.0
pillow==11.1.0
pip==25.1
platformdirs==4.3.7
prompt-toolkit==3.0.43
protobuf==6.31.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
Pygments==2.19.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
setuptools==72.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
stack-data==0.2.0
starlette==0.47.1
sympy==1.14.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typeguard==4.4.3
typer==0.16.0
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
unicodedata2==15.1.0
urllib3==2.3.0
uvicorn==0.34.3
wcwidth==0.2.5
wheel==0.45.1
word2number==1.1
yacs==0.1.8
